\name{AnExample-package}
\alias{AnExample-package}
\alias{AnExample}
\docType{package}
\title{
\packageTitle{AnExample}
}
\description{
\packageDescription{AnExample}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{AnExample}

Maintainer: \packageMaintainer{AnExample}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = AnExample)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{AnExample}}
%\section{Documentation Index}{\packageIndices{AnExample}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
