% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.example.na.R
\docType{data}
\name{d.example.na}
\alias{d.example.na}
\title{Test data for examples in the summary_table function.}
\format{
`d.example.na`
\describe{
A data frame with 45 rows and 19 columns:
  \item{v1}{A character variable. Three levels (Geo 1, Geo 2, Geo 3)}
  \item{v2}{A character variable. Three levels (Other Char 1, Other Char 2, Other Char 3)}
  \item{v3}{A character variable. Three levels (Other Char 1, Other Char 2, Other Char 3)}
  \item{v_pop}{An integer variable. The example population count.}
}
}
\usage{
d.example.na
}
\description{
Test data for examples in the summary_table function.
}
\keyword{datasets}
