% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.example}
\alias{d.example}
\title{Test data for examples in the summary_table function.}
\format{
`d.example`
\describe{
A data frame with 45 rows and 19 columns:
  \item{v1}{A character variable. Three levels (Geo 1, Geo 2, Geo 3)}
  \item{v2}{A character variable. Three levels (Char 1, Char 2, Char 3)}
  \item{v3}{A character variable. Three levels (Other Char 1, Other Char 2, Other Char 3)}
  \item{v4}{A character variable. Two levels (Y, N)}
  \item{v5}{A character variable. Two levels (Y, N)}
  \item{v_pop}{An integer variable. The example population count.}
}
}
\usage{
d.example
}
\description{
Test data for examples in the summary_table function.
}
\keyword{datasets}
