% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_ward}
\alias{d_ward}
\title{Ward Test data for examples in the \code{dc_mapr} function.}
\format{
\code{d_ward}
\describe{
A data frame with 500 rows and 7 columns:
\item{id}{An integer. The unique id}.
\item{ward}{A character variable. The DC ward_ Eight levels (Ward 1:8)}
\item{age}{A numeric variable. The age of the observation.}
\item{cat}{A character variable. Three levels (Cat 1:3)}
\item{bin}{A character variable. Two levels ("Binary 1" "Binary 2")}
\item{bin_other}{An integer variable. Two levels (1, 0)}
\item{ward_pop}{An integer variable. The ward population count.}
}
}
\usage{
d_ward
}
\description{
Ward Test data for examples in the \code{dc_mapr} function.
}
\keyword{datasets}
