% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMADtools-package.R
\docType{package}
\name{DMADtools-package}
\alias{DMADtools}
\alias{DMADtools-package}
\title{DMADtools: Functions and tools for health data analysis}
\description{
Tools for analyzing health and vital records data. Tools for tabulation.
}
\keyword{internal}
