% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{mar2_find}
\alias{mar2_find}
\title{Extract Address Features from a Vector of Washington, D.C. Addresses}
\usage{
mar2_find( address, field = "WARD" )
}
\arguments{
\item{address}{A string vector of addresses.}

\item{field}{A string. The field to return. Must be a column from the \code{tibble} returned by \code{tidymar::find_location()}. See \url{https://github.com/hersh-gupta/tidymar} for a list of fields that can be extracted. The default field is "WARD".}
}
\value{
A character vector with the field requested.
}
\description{
A vectorized function (a wrapper for \code{tidymar::find_location()} ) to extract address features (e.g., ward, census tract, ZIP code, etc.).
}
\details{
This wrapper function uses the \code{tidymar::find_location()} to extract a field from the \code{tibble} returned after querying the District of Columbia’s Master Address Repository (MAR) via the \code{tidymar::find_location()} interface.

Please note that the District of Columbia address should contain the quadrant (e.g., NW, SW, etc.). Otherwise, the address will return as invalid.
}
\section{Citation for tidymar::find_location():}{
Gupta, H. (n.d.). tidymar: An R interface to DC’s Master Address Repository (R package). GitHub. Retrieved from https://github.com/hersh-gupta/tidymar. Accessed on August 26, 2025.
}

\examples{

mar2_find( address = "1600 Pennsylvania Ave NW", field = "WARD" )

mar2_find( address = "1600 Pennsylvania Ave NW", field = "ANC" )

}
