% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{mar2_safe}
\alias{mar2_safe}
\title{Query the DC MAR with MAR2 geocoder.}
\usage{
mar2_safe( address )
}
\arguments{
\item{address}{A string vector of addresses.}
}
\value{
A list of tibbles, with the returns from each queried address.
}
\description{
A type-safe function to query the DC MAR (a wrapper for
\code{tidymar::find_location()} ).
}
\details{
This wrapper function uses the \code{tidymar::find_location()} to validate whether an address exists in the District of Columbia’s Master Address Repository (MAR). Please note that the District of Columbia address should contain the quadrant (e.g., NW, SW, etc.). Otherwise, the address will return as invalid.
}
\section{Citation for tidymar::find_location():}{
Gupta, H. (n.d.). tidymar: An R interface to DC’s Master Address Repository (R package). GitHub. Retrieved from https://github.com/hersh-gupta/tidymar. Accessed on August 26, 2025.
}

\examples{

mar2_safe( address = "1600 Pennsylvania Ave NW" )

}
