% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{validate_address_dc}
\alias{validate_address_dc}
\title{Validate a vector of Washington, D.C. Addresses}
\usage{
validate_address_dc( address )
}
\arguments{
\item{address}{A string vector of addresses.}
}
\value{
An integer vector is returned. If the address was detected in the District of Columbia’s MAR, then a \code{1} is returned. If the address was not detected, a \code{0} is returned.
}
\description{
A vectorized function to validate dc addresses (a wrapper for
\code{tidymar::find_location()} ).
}
\details{
This wrapper function uses the \code{tidymar::find_location()} to validate whether an address exists in the District of Columbia’s Master Address Repository (MAR). Please note that the District of Columbia address should contain the quadrant (e.g., NW, SW, etc.). Otherwise, the address will return as invalid.
}
\section{Citation for tidymar::find_location():}{
Gupta, H. (n.d.). tidymar: An R interface to DC’s Master Address Repository (R package). GitHub. Retrieved from https://github.com/hersh-gupta/tidymar. Accessed on August 26, 2025.
}

\examples{

validate_address_dc( address = "1600 Pennsylvania Ave NW" )

}
