% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_mapr.R
\name{dc_mapr}
\alias{dc_mapr}
\title{Generate a plain or choropleth map of Washington, D.C.}
\usage{
dc_mapr(d, geo, id)
}
\arguments{
\item{d}{A data frame or tibble with the attribute data used to fill the shapes.}

\item{geo}{One of "ward", "zcta", "tract', or "block". The geography that is desired.}

\item{id}{Column name in `d` that contains}
}
\description{
Create a "nice" publication-level map of Washington, D.C. by ward (2012 or 2022 redistricting),
Zip-Code Tabulation Area (ZCTA), Census tract, or block.
}
\details{
Attribute data in `d` (see below) should be organized by ward, ZIP-code tabulation area (ZCTA), Census tract, or block.
}
