% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_mapr.R
\name{dc_mapr}
\alias{dc_mapr}
\title{Generate a plain or choropleth map of Washington, D.C.}
\usage{
dc_mapr(
  d,
  geo,
  var,
  id,
  bypass = FALSE,
  metric = "percent",
  pop.var = NULL,
  percentages.rel = "geo",
  per = 1000,
  count.supp = NULL,
  rate.supp = NULL,
  colorbar.bins = 5,
  colorbar.round = 1,
  colorbar.high = "#7a1315",
  colorbar.low = "#f4e2d9",
  colorbar.h = 0.6,
  colorbar.w = 10,
  colorbar.position = "bottom",
  colorbar.direction = "horizontal",
  colorbar.name = NULL,
  text.color = "black",
  alt.text.color = "grey",
  font.family = "Calibri Light",
  color.thres = 0.4,
  include.compass = TRUE,
  include.scale = TRUE,
  size.scale.title = 2,
  size.scale.labels = 0.7,
  missing.pattern = "stripe",
  suppressed.pattern = "crosshatch",
  pattern.spacing = 0.02
)
}
\arguments{
\item{d}{A data frame or tibble with the attribute data used to fill the shapes.}

\item{geo}{One of "ward 2022", "ward 2012", "zcta", "tract', or "block". The geography that is desired.}

\item{var}{Column name in \code{d} that contains the variable to be plotted.}

\item{id}{Column name in \code{d} that contains the ward, census block, or ZCTA identifier.}

\item{bypass}{A logical. If data supplied in \code{d} is already in aggregate form ready for mapping, then the function will only map the values supplied. Default is FALSE.}

\item{metric}{A vector of string(s). One of: "percent", "mean", "median", "rate", "count". Default if "percent"}

\item{pop.var}{A string or \code{NULL}. Variable in \code{d} with population count data for computing rates if "rate" is selected in \code{metric}. This is normally an aggregate variable. Default is NULL.}

\item{percentages.rel}{A string. One of "dc" or "geo". If \code{metric} is "percent", then this is used to specify denominator for the percentage. "dc" uses \code{nrow(d)} as the denominator while the "geo" option uses geography-specific (e.g., ward- or tract-specific) totals. Default is "geo".}

\item{per}{A numeric. The numerator "per" value to use for computing rates. Default is per \code{1000} (e.g., 25 births per 1,000 individuals).}

\item{count.supp}{an integer or  \code{NULL} if suppression of counts below a certain threshold is desired, specify the integer value. Any cells with counts less than or equal to \code{count.supp} will be suppressed. default is \code{NULL}.}

\item{rate.supp}{An integer or  \code{NULL} if suppression of rates where the count the rate is based off is less than some threshold. Any cells with rates based on counts less than or equal to \code{rate.supp} will be suppressed. default is \code{NULL}.}

\item{colorbar.bins}{An integer. Number of bins for the fill colorbar. Default is 5.}

\item{colorbar.round}{A multiple of 10. Used for rounding colorbar labels. Default is 1.}

\item{colorbar.high}{A Hex color code for the fill gradient. The value used for the "high" end of the scale. Default is "#7a1315".}

\item{colorbar.low}{A Hex color code for the fill gradient. The value used for the "low" end of the scale. Default is "#f4e2d9".}

\item{colorbar.h}{Colorbar height. Default is 0.6.}

\item{colorbar.w}{Colorbar width Default is 10.}

\item{colorbar.position}{Colorbar label position Default is "bottom".}

\item{colorbar.direction}{Colorbar direction. Default is "horizontal".}

\item{colorbar.name}{A string. Name to show beside the colorbar. Default is NULL (i.e., no name displayed).}

\item{text.color}{Text color. Default is "black".}

\item{alt.text.color}{An alternate text color, for when fill color and text color are too similar. Default is "grey".}

\item{font.family}{A string with the font desired for text on the map. Default is "Calibri Light".}

\item{color.thres}{A numeric. A threshold value (in \link{0,1}) for judging similarity of fill color and text color. Default is 0.4}

\item{include.compass}{A logical. Include reference compass rose on map? Default is TRUE.}

\item{include.scale}{A logical. Include reference scale rose on map? Default is TRUE.}

\item{size.scale.title}{A numeric. Scaling parameter for title sizing on map.}

\item{size.scale.labels}{A numeric. Scaling parameter for label sizing on map.}

\item{missing.pattern}{A theme from \code{ggpattern} for polygons with missing data. Default is "stripe". See \href{https://coolbutuseless.github.io/package/ggpattern/}{ggpattern}.}

\item{suppressed.pattern}{A theme from \code{ggpattern} for polygons with suppressed data. Default is "weave". See \href{https://coolbutuseless.github.io/package/ggpattern/}{ggpattern}.}

\item{pattern.spacing}{A numeric. Passed to \code{pattern_spacing} argument of\code{ggpattern::geom_pattern()}.}
}
\description{
Create a "nice" publication-level map of Washington, D.C. by ward (2012 or 2022 redistricting),
Zip-Code Tabulation Area (ZCTA), Census tract, or block.
}
\details{
Attribute data in \code{d} (see below) should be organized by ward, ZIP-code tabulation area (ZCTA), census tract, or block.
The following DC geographies can be plotted using this function.\tabular{lcl}{
   Geography \tab \code{geo} call \tab Accepted formats in \code{id} \cr
   2022 Wards \tab "ward 2022" \tab i. "Ward 1"..."Ward 8"\if{html}{\out{<br>}}ii. "ward 1"..."ward 8"\if{html}{\out{<br>}}iii. "One"..."Eight"\if{html}{\out{<br>}}iv. "one"..."eight" \cr
}
}
\section{Citation for Shapefiles}{
U.S. Census Bureau, “2024 TIGER/Line Shapefiles” 2024,
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html},
accessed on May 29, 2025.
}

