% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Generate "Nice" \code{flextable} From Input Data}
\usage{
summary_table( d, var1, var2 = NULL, table.grouping = NULL, pop.var = NULL,
add.summary.row = TRUE, summary.row.name = NULL, add.summary.col = TRUE, font.family = "Calibri Light",
digs.perc = 1, digs.rate = 2, summary.col.name = NULL, summary.col.pos = "front", order.rows = NULL, order.cols = NULL, 
order.groups = NULL, foot.lines = c(), table.title = NULL, metric = c( "count", "rate" ), nm.var1 = NULL, 
count.supp = NULL, remove.cols = NULL, rate.supp = NULL, count.supp.symbol = "--", rate.supp.symbol = "*", 
per = 1000, NAs.footnote = FALSE, percentages.rel = "var1", include.percent.sign = TRUE, 
row.variable.labels = "default", col.variable.labels = "default", row.variable.col = "#eed8a4" )
}
\arguments{
\item{d}{A data frame or tibble. Data must be the exact subset of data that needs to analyzed for the table generation.}

\item{var1}{A string. Name of first variable to stratify on (as it appears in the input data, \code{d}). Cannot be  \code{NULL}.}

\item{var2}{A string. Name of second variable to stratify on (as it appears in the input data, \code{d}). Can be \code{NULL} if only one variable is required and there is no cross-tabulation or if \code{table.grouping} and \code{var1} are specified for a cross-tabulation instead.}

\item{table.grouping}{A string or \code{NULL} (if no grouping in table is desired). Name of variable to create grouped output table with (creates grouping rows in the table). Can be 1 variable at maximum.}

\item{pop.var}{A string or \code{NULL}. Variable in \code{d} with population count data for computing rates if "rate" is selected in \code{metric}. This is normally an aggregate variable.}

\item{add.summary.row}{A logical. Add a row with unaggregated results (based on \code{var1}). \code{TRUE} is default and the name given to the row is "Summary Row". If another name is desired, it can be specified with \code{summary.row.name} and its order/position can be specified with the \code{order.rows} argument.}

\item{summary.row.name}{A string. Name to appear for summary row label. Default is "Summary Row". This option is only relevant if \code{add.summary.row == TRUE}.}

\item{add.summary.col}{A logical. Add a row with unaggregated results (based on \code{var2}). Note that this argument is inconsequential if \code{var2 == NULL}.}

\item{font.family}{A string with the font desired, to be applied to all sections of the table.}

\item{digs.perc}{An integer. Digits to round percents to.}

\item{digs.rate}{An integer. Digits to round rates to.}

\item{summary.col.name}{A string. Name to appear for summary column header. Default is "All". This option is only relevant if \code{add.summary.col == TRUE}.}

\item{summary.col.pos}{One of "front", "end", or an integer >= 0 and < the number of all levels of the categories in \code{var2} + 1.  Only a relevant argument if \code{length( var2 ) > 1} & \code{add.summary.col} is \code{TRUE}.}

\item{order.rows}{A vector with the levels in custom order of the stratifying variable in the rows(typically \code{var1} but can also be \code{table.grouping} depending on how the function is specified). Entries must match levels of stratifying variable exactly. If there are any levels of the row variable missing, they are omitted from the final table. Note that any rows removed from the final table do not remove that level from the data itself. If desired, that must be done outside the function with a data step. If multiple variables are listed in \code{var1}, then a named list with the variables specified in \code{var1} as the entry names and a character string with the desired order of each of those variable levels should be specified.}

\item{order.cols}{A vector with the levels in custom order of the stratifying variable in the columns (typically \code{var2} but can also be \code{table.grouping} depending on how the function is specified). Must match the columns in the final table including the summary column name if \code{add.summary.col == TRUE} and \code{summary.col.name} is specified. If there are any levels of the column variable missing, they are omitted from the final table.}

\item{order.groups}{A vector with the levels in custom order of the grouping variable in the columns. Must match the grouping variable (\code{table.grouping}) names in the final table. if there are any levels of the grouping variable missing, they are omitted from the final table. the orders of the rows within the group are preserved as indicated in \code{order.rows}. this argument is only valid if \code{!is.null( var2 ) & !is.null( table.grouping )}.}

\item{foot.lines}{A vector with entries (strings) corresponding to each footline to be added to the table.}

\item{table.title}{A string. Name of table to display as the title. \code{NULL} is default for no title in the table.}

\item{metric}{A vector of string(s). One or two of: "percent", "rate", "count". Can select up to two options to plot at maximum.}

\item{nm.var1}{A string. Name of \code{var1} that you would like to appear in the table. \code{NULL} is default, which prompts the column name as it is written in the input dataset, \code{d}.}

\item{count.supp}{an integer or  \code{NULL} if suppression of counts below a certain threshold is desired, specify the integer value. Any cells with counts less than or equal to \code{count.supp} will be suppressed. default is \code{NULL}.}

\item{remove.cols}{A string. The string is fed to \verb{dplyr::select( -contains( remove.cols} ) )` to remove undesired columns from the final table. Should be based off of columns seen in final generated table.}

\item{rate.supp}{An integer or  \code{NULL} if suppression of rates where the count the rate is based off is less than some threshold. Any cells with rates based on counts less than or equal to \code{rate.supp} will be suppressed. default is \code{NULL}.}

\item{count.supp.symbol}{A string. The character to display in the table where the count is suppressed.}

\item{rate.supp.symbol}{A string. The character to display in the table where the rate is suppressed.}

\item{per}{A numeric. The numerator "per" value to use for computing rates. Default is per \code{1000} (e.g., 25 births per 1,000 individuals).}

\item{NAs.footnote}{A logical. Include footnotes detailing number of missing values in the dataset based on \code{var1} and \code{var2}?}

\item{percentages.rel}{A string. One of "var1" or "var2" or "table.grouping". Is the variables with which percentages should be calculated with respect to. \code{table.grouping} can only be specified if only \code{var1} is specified and \code{var2 == NULL}.}

\item{include.percent.sign}{A logical. Include percent (\%) sign in computed percent columns?}

\item{row.variable.labels}{"default", "none" or a named list with the variables specified in \code{var1} as the entry names and a character string with the desired label for that variable in the rows.  If "default", the default variable names (i.e., those listed in \code{var1} are printed). If "none", no labels are printed. Default is "default".}

\item{col.variable.labels}{"default", "none" or a named list with the variables specified in \code{var2} as the entry names and a character string with the desired label for that variable in the columns  If "default", the default variable names (i.e., those listed in \code{var2} are printed). If "none", no labels are printed. Default is "default".}

\item{row.variable.col}{A string Hex code or color code for the background in the label row or "none". This argument is only relevant if \code{length( var1 ) > 1} (i.e., multiple variables are desired in the rows of the table). Default is "#eed8a4". If "none" the default \code{flextable::theme_zebra} theme is used for that row.}
}
\value{
Object of class \code{list} containing the following elements:

\code{flextable}: An object of class \code{flextable} that produces the "nice" flextable.

\code{frame}: An object of class \code{data.frame} that contains the dataset fed to \code{flextable} for final tabulation.

NOTE: In the final output table (\code{flextable}), levels of \code{var1} appear in the rows and levels of \code{var2} appear in the columns if only \code{var1} and \code{var2} are specified. If \code{var1} and \code{table.grouping}
are only specified, then levels of \code{var1} appear in the columns and levels of \code{table.grouping} appear in the rows. If \code{var1}, \code{var2}, and \code{table.grouping} are specified, then levels of \code{var1}
appear in the rows, levels of \code{var2} appear in the columns, and levels of \code{table.grouping} appear as nested rows that group the table.
}
\description{
Create a "nice" publication-level table by cross-tabulating up to three variables at once.
}
\details{
Employs \code{flextable} to generate a "nice" publication-level table that can be easily exported or included in R Rmarkdown (.RMD) files.
}
\examples{

library( DMADtools )

summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               var2 = "v2",
               percentages.rel = "var1" )


# alternate percentages.rel argument
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               var2 = "v2",
               percentages.rel = "var2" )

# specify var1 and table.grouping instead of var1 and var2, 
# note, the change in percentages.rel            
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               table.grouping = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               percentages.rel = "table.grouping" )

# rename summary row and column and shuffle column and row orders
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               table.grouping = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               summary.col.name = "All Geos", # change name of summary column
               summary.row.name = "All Char", # change name of summary row
               order.cols = c( "Geo 1", 
               "Geo 2", "All Geos", "Geo 3" ), # reorder columns in output table
               order.rows = c( "Char 1", 
               "Char 2", "All Char", "Char 3" ), # reorder columns in output table
               percentages.rel = "table.grouping" )

# same call as previous, but use order.cols and order.rows to remove some 
# rows and columns from final table
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               table.grouping = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               summary.col.name = "All Geos", # change name of summary column
               summary.row.name = "All Char", # change name of summary row
               order.cols = c( "Geo 1", 
               "All Geos", "Geo 3" ), # reorder columns in output table
               order.rows = c( "Char 1", 
               "Char 2", "All Char" ), # reorder columns in output table
               percentages.rel = "table.grouping" )

# use remove.cols instead of order.cols argument for removing an undesired column from final table
summary_table( d = d_example,
               metric = c( "count", "rate" ),
               var1 = "v1",
               var2 = "v2",
               add.summary.row =TRUE,
               add.summary.col = TRUE,
               remove.cols = "Char 3",
               percentages.rel = "var2",
               pop.var = "v_pop" )

# rename var1 column header
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               nm.var1 = "Variable 1", # for relabeling the header
               table.grouping = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               summary.col.name = "All Geos", # change name of summary column
               summary.row.name = "All Char", # change name of summary row
               order.cols = c( "Geo 1", 
               "All Geos", "Geo 3" ), # reorder columns in output table
               order.rows = c( "Char 1",
               "Char 2", "All Char" ), # reorder columns in output table
               percentages.rel = "table.grouping" )

# do a 3-way tabulation
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               var2 = "v3",
               table.grouping = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               percentages.rel = "var2" )

# use order.groups argument to remove some grouping categories
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = "v1",
               var2 = "v3",
               table.grouping = "v2",
               order.groups = c( "Char 2", "Char 1" ),
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               percentages.rel = "var2" )

# example with rate computation and add.summary.row and add.summary.col are TRUE
summary_table( d = d_example,
               metric = c( "count", "rate" ),
               var1 = "v1",
               var2 = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               percentages.rel = "var2",
               pop.var = "v_pop" )

# foot.lines argument and NAs.footnote

summary_table( d = d_example_na,
               metric = c( "count", "percent" ),
               var1 = "v1",
               var2 = "v2",
               nm.var1 = "Variable 1",
               add.summary.row =TRUE,
               add.summary.col = TRUE,
               foot.lines = c("This is the first footer line",
                              "This is the second footer line" ), 
               remove.cols = "Char 3",
               percentages.rel = "var2",
               NAs.footnote = TRUE, # produces the footnote of excluded observations
               pop.var = "v_pop" ) 

# add rate suppression rule
summary_table( d = d_example_na,
               metric = c( "count", "rate" ),
               var1 = "v1",
               var2 = "v2",
               nm.var1 = "Variable 1",
               add.summary.row =TRUE,
               add.summary.col = TRUE,
               foot.lines = c("This is the first footer line",
                              "This is the second footer line" ),
               remove.cols = "Char 3",
               percentages.rel = "var2",
               count.supp = 50, # suppress values/counts in table <= 50
               rate.supp = 50, # suppress rate calculation when counts are <= 50
               NAs.footnote = TRUE, # produces the footnote of excluded observations
               pop.var = "v_pop" ) 

# give it a title
summary_table( d = d_example_na,
               metric = c( "count", "rate" ),
               var1 = "v1",
               var2 = "v2",
               nm.var1 = "Variable 1",
               add.summary.row =TRUE,
               add.summary.col = TRUE,
               foot.lines = c("This is the first footer line",
                              "This is the second footer line" ), 
               remove.cols = "Char 3",
               percentages.rel = "var2",
               count.supp = 50, # suppress values/counts in table <= 50
               rate.supp = 50, # suppress rate calculation when counts are <= 50
               NAs.footnote = TRUE, # produces the foornote of excluded observations
               pop.var = "v_pop",
               table.title = "This is Table 1" ) # title
               
# when providing more than 1 variable for the rows
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = c( "v1", "v3" ),
               order.rows = list( v1 = c( "Geo 2", "Geo 3", "Geo 1" ),
                                  v3 = c( "Other Char 2", "Other Char 3" )),
               var2 = "v2",
               add.summary.row = FALSE,
               add.summary.col = FALSE,
               rate.supp = 5,
               count.supp = 5,
               percentages.rel = "var2",
               pop.var = "v_pop" )
        
# use of  `row.variable.labels`     
summary_table( d = d_example,
               metric = c( "count", "percent" ),
               var1 = c( "v1", "v3" ),
               order.rows = list( v1 = c( "Geo 2", "Geo 3", "Geo 1" ),
                                  v3 = c( "Other Char 2", "Other Char 3" )),
               var2 = "v2",
               add.summary.row = TRUE,
               add.summary.col = TRUE,
               rate.supp = 5,
               count.supp = 5,
               percentages.rel = "var2",
               row.variable.labels = list( v1 = "var1",
                                           v3 = "var3"),
               pop.var = "v_pop" )

}
